create ControlsIndicatorManager, bring back indicator on video click
